add Square and Pentagon file, move functions and import them, link randomColor correctly
