+++ js-structure/js-structure_shapes/components/Circle/Circle.js
@@ -1,3 +1,5 @@
+import { getRandomColor } from "../../utils/randomColor.js";
+
 export default function Circle() {
     const circle = document.createElement("div");
     circle.classList.add("circle");
